refactor: move swiper styles to a shared space
