+++ src/components/shared_styles/swiper-styles.ts
